add  Controllers and services

+++ src/main/java/org/demointernetshop/dto/product/CategoryWithPricesDto.java
@@ -5,6 +5,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -13,9 +15,9 @@ public class CategoryWithPricesDto {
     @Schema(description = "Category identifier", example = "1")
     private Integer id;
     @Schema(description = "minPrice", example = "201")
-    private Integer minPrice;
+    private BigDecimal minPrice;
     @Schema(description = "maxPrice", example = "10000")
-    private Integer maxPrice;
+    private BigDecimal maxPrice;
     @Schema(description = "Name of the Category", example = "smartphones")
     private String categoryName;
 }
